+++ src/commands/index.ts
@@ -1,4 +1,6 @@
 export { CreateTableParams } from "./create-table.js";
+export * from "./expressions/index.js";
 export { GetItemParams } from "./get-item.js";
 export { PutItemParams } from "./put-item.js";
+export { QueryParams } from "./query.js";
 export { WriteTransactionParams } from "./write-transaction.js";